Correccion visualizacion al mostrar paso a paso de base 10 a base destino (Revisar if comentado). Mensaje de error argn ivalido mas detallado.

+++ impresion.h
@@ -5,9 +5,9 @@ void papDecimalOBaseT10Base(int * total, int * n, int * OBase, int * exp);
 
 void papDecimal10BaseTDBase(int * n, int * b, int * r);
 
-void papFractionaryOBaseT10Base(int * total, int * n, int * OBase, int * exp);
+void papFractionaryOBaseT10Base(float* total, int * n, int * OBase, int * exp);
 
-void papFractionary10BaseTDBase(int * TBase, int * n, int * digit);
+void papFractionary10BaseTDBase(int * TBase, float * n, int * digit);
 
 void mostrarResultadoFinal(char * resultadoEntero, char * resultadoFraccional);
